requirements.txt
Flask==2.2.3
flask-restx==1.0.6
Flask-Cors==3.0.10
gunicorn==20.1.0
pytest==6.2.5
python-dotenv==0.19.2
Werkzeug==2.2.2
pyOpenSSL==22.0.0
requests~=2.25.1
pyhsm~=1.2.1.post1
setuptools~=67.4.0